pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "discovery-engine"
version = "0.1.0"
description = "Distributed scientific paper extraction for cross-domain discovery"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "Discovery Engine Contributors" },
]
keywords = ["science", "discovery", "extraction", "cross-domain", "knowledge-graph"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Topic :: Scientific/Engineering",
]

dependencies = [
    "httpx>=0.25",
]

[project.optional-dependencies]
anthropic = ["anthropic>=0.40"]
openai = ["openai>=1.0"]
gemini = ["google-generativeai>=0.8"]
all = [
    "anthropic>=0.40",
    "openai>=1.0",
    "google-generativeai>=0.8",
]
hf = [
    "datasets>=2.0",
    "huggingface_hub>=0.20",
]
dev = [
    "pytest>=8.0",
    "ruff>=0.4",
]

[project.scripts]
discovery = "discovery.cli:main"

[project.urls]
Repository = "https://github.com/discovery-engine/discovery-engine"
Documentation = "https://github.com/discovery-engine/discovery-engine/blob/main/DESIGN.md"

[tool.hatch.build.targets.wheel]
packages = ["discovery"]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.pytest.ini_options]
testpaths = ["tests"]
